Создала отдельный сценарий для общих функций
